Bugfix creating/updating detachments

+++ app/controllers/api/detachments_controller.rb
@@ -1,6 +1,6 @@
 class Api::DetachmentsController < ApplicationController
 	def create
-		detachment = Detachment.create detachment_params
+		detachment = Detachment.create params[:detachment]
 		render json: detachment.to_json
 	end
 
@@ -8,34 +8,36 @@ class Api::DetachmentsController < ApplicationController
 		detachment = Detachment.find params[:id]
 		detachment.update params[:detachment]
 
-		params[:detachment_units].each do |detachment_unit_params|
-			detachment_unit_slots = detachment_unit_params.delete(:detachment_unit_slots)
-			(detachment_unit_slots || []).each do |slot_params|
-				slot = DetachmentUnitSlot.find_by id: slot_params[:id]
-				if slot
-					slot.update slot_params
+		if params[:detachment_units]
+			params[:detachment_units].each do |detachment_unit_params|
+				detachment_unit_slots = detachment_unit_params.delete(:detachment_unit_slots)
+				(detachment_unit_slots || []).each do |slot_params|
+					slot = DetachmentUnitSlot.find_by id: slot_params[:id]
+					if slot
+						slot.update slot_params
+					else
+						DetachmentUnitSlot.create slot_params
+					end
+				end
+
+				d_unit = DetachmentUnit.find_by id: detachment_unit_params[:id]
+				if d_unit
+					d_unit.update detachment_unit_params
 				else
-					DetachmentUnitSlot.create slot_params
+					DetachmentUnit.create detachment_unit_params
 				end
 			end
 
-			d_unit = DetachmentUnit.find_by id: detachment_unit_params[:id]
-			if d_unit
-				d_unit.update detachment_unit_params
+			unit_ids = params[:detachment_units].map{|unit| unit[:unit_id]}
+			if unit_ids.size > 0
+				DetachmentUnit.connection.execute "DELETE from detachment_units
+				WHERE detachment_id = #{params[:id]}
+				AND unit_id NOT IN (#{unit_ids.join(', ')});"
 			else
-				DetachmentUnit.create detachment_unit_params
+				# Case for removing last unit from detachment
+				DetachmentUnit.connection.execute "DELETE from detachment_units
+				WHERE detachment_id = #{params[:id]};"
 			end
-		end
-
-		unit_ids = params[:detachment_units].map{|unit| unit[:unit_id]}
-		if unit_ids.size > 0
-			DetachmentUnit.connection.execute "DELETE from detachment_units
-			WHERE detachment_id = #{params[:id]}
-			AND unit_id NOT IN (#{unit_ids.join(', ')});"
-		else
-			# Case for removing last unit from detachment
-			DetachmentUnit.connection.execute "DELETE from detachment_units
-			WHERE detachment_id = #{params[:id]};"
 		end
 
 		render json: detachment.to_json
